backend/requirements.txt
# === Core Framework ===
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.5.2

# === Authentication ===
python-jose[cryptography]==3.3.0
authlib==1.3.2
httpx==0.27.2
itsdangerous==2.2.0

# === LangChain + LangGraph ===
langchain==0.3.14
langchain-community==0.3.14
langchain-openai==0.3.0
langchain-text-splitters==0.3.4
langgraph==0.2.60

# === Embeddings (Jina AI) ===
langchain-community>=0.3.14

# === Vector DB (Qdrant) ===
qdrant-client==1.12.1

# === Document Processing ===
pymupdf==1.25.1

# === PII Protection (Presidio) ===
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355
spacy==3.7.6
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl

# === Database ===
pymongo==4.10.1

# === Supabase ===
supabase==2.11.0

# === Redis (Cache + Analytics) ===
upstash-redis==1.1.0

# === Resilience ===
pybreaker==1.2.0

# === Observability ===
langfuse>=2.0.0

# === Rate Limiting ===
slowapi==0.1.9

# === Utilities ===
python-dotenv==1.0.1
python-multipart==0.0.12
numpy<2.0.0

# === Privacy & Security ===
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355
# Direct download for deployment (Render/Koyeb)
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
